Update version information representation
The recent change of the version information representation was
unfortunately problematic for some downstream tools like `rcatmaid`.
They rely on the commit counter to be available to compare versions
between releases. Therefore, the version representation now uses a
format similar to `git describe` again:

  <base-version>-<commit-counter>-<commit>

The <base-version> is taken from BASE_VERSION field, the commit counter
is computed based on the distance of the current commit to the
BASE_COMMIT field id. This new field is populated by the release script.
The <commit> reference is the current commit. Both <commit> and
<commit-counter> are populated from `git describe`, if it is available.
If not, the files /home/git-commit and /home/git-base-version are
expected to be present (Docker context). If they  are not available,
the BASE_VERSION variable is returned together with the suffix
"-unknown".